refactor useLink into a module
